A63.cpp
The program gets the number of inputs from the main program and then run the getSum function which adds all of the inputs together when it asks for each input. Then the program determines which program is the min and the max in order to subtract them from the total. This value is then output onto the console.